tokenAdmin.getInflationRate
